refactors kubeapiserver webhook utility to be generic
